requirements.txt
Pillow
opencv-python
PyQt5
pyyaml
tqdm